Add phpdocs.
Also added a composer script entry so the docs can be generated using
'./composer.phar docs'.

NOTE: Phpdocumentor does not fully support PHP7 yet, so the docs are
not generated for anything other than the exceptions at the moment.
It looks like they're planning a release in the next few days that
will.  Leaving ticket #22 open until that happens because until then
we won't know whether everything is rendering properly!